+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/myMealButton/MyMealButtonEmptyView.kt
@@ -1,0 +1,10 @@
+package com.miam.kmm_miam_sdk.android.ui.components.myMealButton
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MyMealButtonEmptyView() {
+    //TODO add template
+    Box {}
+}