+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/myMealButton/customization/MyMealButtonText.kt
@@ -1,0 +1,6 @@
+package com.miam.kmm_miam_sdk.android.ui.components.myMealButton.customization
+
+object MyMealButtonText {
+    var singleMealText = "1 repas ajouté"
+    var multipleMealText = " repas ajoutés"
+}